# rizomai/openapi/rizomai.yaml
...
openapi: 3.1.0
info:
  title: RIZOMAI API
  version: 0.1.0
  description: |
    Plataforma API-first de gestão de redes sociais — "um post → todas as redes".

    Contrato v0.1 (scaffold). Cobre os recursos centrais: Profiles, Social
    Accounts (OAuth broker), Posts (fan-out por plataforma), Media (upload
    direto ≤25MB / presign até 5GB) e Webhooks (HMAC-SHA256).

    Autenticação: toda chamada usa `Authorization: Bearer sk_...` (API key).
    Mutações (`POST`/`PUT`) aceitam o header `Idempotency-Key` (UUID): a mesma
    chave dentro da janela de 24h retorna a mesma resposta.

    Erros: envelope `{ "code", "error", "details" }` em todo 4xx/5xx.
    `code` é machine-readable em UPPER_SNAKE (ex.: RATE_LIMITED).

    Documentação dos ADRs: docs/projetos/rizomai/adr/ no workspace do Cosca.

  contact:
    name: RIZOMAI
    url: https://rizomai.app
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: https://api.rizomai.app
    description: Produção
...
security:
  - bearerAuth: []
paths:
...
  /v1/profiles:
    get:
      operationId: listProfiles
      tags: [Profiles]
      summary: Lista perfis do tenant (paginado)
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Lista de perfis
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
    post:
      operationId: createProfile
      tags: [Profiles]
      summary: Cria um perfil
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileCreate'
            example:
              name: Minha Agência
      responses:
        '201':
          description: Perfil criado
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/Profile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/profiles/{id}:
    parameters:
      - $ref: '#/components/parameters/ProfileId'
    get:
      operationId: getProfile
      tags: [Profiles]
      summary: Obtém um perfil
      responses:
        '200':
          description: Perfil encontrado
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/Profile'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
    put:
      operationId: updateProfile
      tags: [Profiles]
      summary: Atualiza um perfil
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileUpdate'
      responses:
        '200':
          description: Perfil atualizado
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/Profile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
    delete:
      operationId: deleteProfile
      tags: [Profiles]
      summary: Remove um perfil
      responses:
        '204':
          description: Perfil removido
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'

  # ---------------------------------------------------------------------------
  # Social Accounts (OAuth broker — ADR-006)
  # ---------------------------------------------------------------------------
...
  /v1/posts:
    get:
      operationId: listPosts
      tags: [Posts]
      summary: Lista posts (paginado)
      description: Filtra por profileId e/ou status agregado.
      parameters:
        - $ref: '#/components/parameters/ProfileIdQuery'
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/PostStatus'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Lista de posts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
    post:
      operationId: createPost
      tags: [Posts]
      summary: Cria um post com fan-out para N plataformas
      description: |
        Cria o Post + N PostTargets atomicamente (transação — ADR-002/003).
        Idempotência em 2 camadas (ADR-005 §1.3): header `Idempotency-Key` E
        content-hash dedup ((team, content+media+targets) em 24h → 409 com
        existingPostId).
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostCreate'
            example:
              content: Olá mundo
              platforms:
                - platform: x
                  accountId: account_123
                - platform: linkedin
                  accountId: account_456
                  platformSpecificData:
                    firstComment: Saiba mais no link
                    visibility: PUBLIC
              mediaUrls:
                - https://cdn.rizomai.app/media_abc
              scheduledFor: 2026-09-14T13:00:00Z
              timezone: America/Sao_Paulo
      responses:
        '201':
          description: Post criado (status agregado scheduled)
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/Post'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/posts/{id}:
    parameters:
      - $ref: '#/components/parameters/PostId'
    get:
      operationId: getPost
      tags: [Posts]
      summary: Obtém um post com o status de cada target
      responses:
        '200':
          description: Post encontrado
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/Post'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
    delete:
      operationId: deletePost
      tags: [Posts]
      summary: Cancela/remove um post
      description: |
        Só permite cancelar posts em status scheduled/publishing. Post já
        publicado total ou parcialmente exige unpublish por target (Fase 2).
      responses:
        '204':
          description: Post removido
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'

  # ---------------------------------------------------------------------------
  # Media (ADR-008)
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API key
      description: |
        API key do tenant no formato `sk_...`. Enviar no header:
        `Authorization: Bearer sk_...` (ADR-006 — o cliente nunca vê tokens de
        rede social, apenas a API key).
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        UUID do cliente para idempotência (ADR-005 §1.3). A mesma chave dentro
        da janela de 24h devolve a mesma resposta (armazenada com o request hash).
      schema:
        type: string
        format: uuid
        example: 3f9b8c2a-1c4e-4f7a-9d5b-0e2a1c3d4e5f
    ProfileId:
      name: id
      in: path
      required: true
      schema:
        type: string
        pattern: '^profile_[a-zA-Z0-9]+$'
        example: profile_abc123
    ProfileIdQuery:
      name: profileId
      in: query
      required: false
      schema:
        type: string
        pattern: '^profile_[a-zA-Z0-9]+$'
        example: profile_abc123
...
    PostId:
      name: id
      in: path
      required: true
      schema:
        type: string
        pattern: '^post_[a-zA-Z0-9]+$'
        example: post_abc123
...
    Page:
      name: page
      in: query
      required: false
      description: Página (1-indexada)
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      required: false
      description: Itens por página (máx. 100)
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    BadRequest:
      description: Requisição malformada (JSON inválido, campos ausentes)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: BAD_REQUEST
            error: Corpo da requisição inválido
            details: {}
    Unauthorized:
      description: API key ausente, inválida ou expirada
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: UNAUTHORIZED
            error: API key inválida
            details: {}
...
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: NOT_FOUND
            error: Recurso não encontrado
            details: { resource: profile, id: profile_abc123 }
    Conflict:
      description: "Conflito de estado (ex.: dedup de post duplicado)"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: CONFLICT
            error: Post duplicado nos últimos 24h
            details: { existingPostId: post_xyz789 }
    ValidationError:
      description: Falha de validação (campo/valor fora do contrato)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: VALIDATION_ERROR
            error: Dados inválidos
            details:
              fields:
                platforms[0].platform: Plataforma não suportada
...
    RateLimited:
      description: |
        Rate limit excedido (por API key/tenant). O middleware envia os headers
        `X-RateLimit-Limit`, `X-RateLimit-Remaining` e `X-RateLimit-Reset`
        (epoch) em TODAS as respostas 2xx/4xx; quando negado, também envia
        `Retry-After` (segundos).
      headers:
        X-RateLimit-Limit:
          description: "Limite de requests por janela (ex.: 60)."
          schema:
            type: integer
        X-RateLimit-Remaining:
          description: Requests restantes na janela atual.
          schema:
            type: integer
        X-RateLimit-Reset:
          description: Epoch (segundos) em que a janela/limite reseta.
          schema:
            type: integer
            format: int64
        Retry-After:
          description: Segundos até o próximo request ser aceito (apenas quando negado).
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: RATE_LIMITED
            error: Rate limit excedido — tente novamente mais tarde
            details: { retryAfterSeconds: 60 }
...
  schemas:
...
    Error:
      type: object
      required: [code, error]
      properties:
        code:
          type: string
          description: Código machine-readable em UPPER_SNAKE.
          enum:
            - BAD_REQUEST
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND
            - VALIDATION_ERROR
            - CONFLICT
            - RATE_LIMITED
            - PAYMENT_REQUIRED
            - NOT_IMPLEMENTED
            - INTERNAL_ERROR
          example: RATE_LIMITED
        error:
          type: string
          description: Mensagem legível por humanos (PT-BR).
          example: Rate limit excedido
        details:
          type: object
          description: "Contexto estruturado (ex.: { fields: {...} }, { existingPostId })."
          additionalProperties: true

    # --- Plataformas (ADR-006: MVP = x, linkedin, telegram; Fase 2+ amplia) ---
    Platform:
      type: string
      description: |
        Rede social alvo. MVP (ADR-006 §1): x, linkedin, telegram.
        Fase 2+: instagram, tiktok, youtube, bluesky, reddit, whatsapp, googlebusiness.
      enum: [x, linkedin, telegram]
      example: x

    # --- Profiles ---
    Profile:
      type: object
      required: [id, name, createdAt, updatedAt]
      properties:
        id:
          type: string
          pattern: '^profile_[a-zA-Z0-9]+$'
          example: profile_abc123
        name:
          type: string
          minLength: 1
          maxLength: 100
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProfileCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
    ProfileUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
    ProfileList:
      type: object
      required: [data, page, limit, total]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Profile'
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer

    # --- Social Accounts (ADR-006) ---
...
    PostStatus:
      type: string
      description: |
        Status agregado DERIVADO dos targets (ADR-007 §1). Nunca editado à mão.
      enum: [scheduled, publishing, published, partial, failed, cancelled]
      example: partial
    TargetStatus:
      type: string
      enum: [pending, scheduled, publishing, published, failed, skipped]
    TargetError:
      type: object
      required: [code]
      properties:
        code:
          type: string
          description: "Código tipado do erro (ex.: invalid_grant, duplicate)."
        error:
          type: string
          description: Mensagem legível.
    PostTarget:
      type: object
      required: [platform, accountId, status]
      properties:
        platform:
          $ref: '#/components/schemas/Platform'
        accountId:
          type: string
          pattern: '^account_[a-zA-Z0-9]+$'
        status:
          $ref: '#/components/schemas/TargetStatus'
        platformSpecificData:
          type: object
          additionalProperties: true
          description: |
            Dados específicos da plataforma (union tipada por plataforma na
            spec — ADR-005 §1.5). JSONB no banco (ADR-002), validado na API.
        publishedUrl:
          type: [string, 'null']
          format: uri
          description: URL pública do post publicado na rede (quando aplicável).
        externalPostId:
          type: [string, 'null']
          description: Id externo do post na plataforma.
        lastError:
          $ref: '#/components/schemas/TargetError'
    PostTargetCreate:
      type: object
      required: [platform, accountId]
      properties:
        platform:
          $ref: '#/components/schemas/Platform'
        accountId:
          type: string
          pattern: '^account_[a-zA-Z0-9]+$'
        platformSpecificData:
          type: object
          additionalProperties: true
    Post:
      type: object
      required: [id, profileId, content, platforms, status, createdAt, updatedAt]
      properties:
        id:
          type: string
          pattern: '^post_[a-zA-Z0-9]+$'
          example: post_abc123
        profileId:
          type: string
          pattern: '^profile_[a-zA-Z0-9]+$'
        content:
          type: string
          minLength: 1
          maxLength: 4000
        mediaUrls:
          type: array
          description: URLs públicas da mídia (de upload direto ou presign — ADR-008).
          items:
            type: string
            format: uri
        platforms:
          type: array
          description: Targets do fan-out (1 por conta/plataforma).
          items:
            $ref: '#/components/schemas/PostTarget'
        scheduledFor:
          type: [string, 'null']
          format: date-time
          description: Momento de publicação (null = publicar já).
        timezone:
          type: string
          default: UTC
          example: America/Sao_Paulo
        status:
          $ref: '#/components/schemas/PostStatus'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PostCreate:
      type: object
      required: [content, platforms]
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 4000
        platforms:
          type: array
          minItems: 1
          maxItems: 50
          items:
            $ref: '#/components/schemas/PostTargetCreate'
        mediaUrls:
          type: array
          items:
            type: string
            format: uri
        scheduledFor:
          type: [string, 'null']
          format: date-time
          description: Momento de publicação (null/omiti do = publicar já).
        timezone:
          type: string
          default: UTC
          description: "IANA (ex.: America/Sao_Paulo) — usado com scheduledFor."
    PostList:
      type: object
      required: [data, page, limit, total]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Post'
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer

    # --- Media (ADR-008) ---
...